# [JENKINS-67344] Poor performance of `UpdateSite.getData`