--- a/src/main/java/net/sf/json/JSONObject.java
+++ b/src/main/java/net/sf/json/JSONObject.java
@@ -943,7 +943,7 @@ public final class JSONObject extends AbstractJSON implements JSON, Map<String, 
             String key;
             Object value;
 
-            if (tokener.matches("null.*")) {
+            if (tokener.startsWith("null")) {
                 fireObjectStartEvent(jsonConfig);
                 fireObjectEndEvent(jsonConfig);
                 return new JSONObject(true);