--- a/src/main/java/net/sf/json/util/JSONTokener.java
+++ b/src/main/java/net/sf/json/util/JSONTokener.java
@@ -111,11 +111,11 @@ public class JSONTokener {
     }
 
     public boolean startsWith(String prefix) {
-        int patternLength = prefix.length();
-        if (length() - this.myIndex < patternLength) {
+        int prefixLength = prefix.length();
+        if (length() - this.myIndex < prefixLength) {
             return false;
         }
-        String str = this.mySource.substring(this.myIndex, this.myIndex + patternLength);
+        String str = this.mySource.substring(this.myIndex, this.myIndex + prefixLength);
         return str.startsWith(prefix);
     }