Profiling showed that while regex compilation was indeed a problem, it wasn't the bottleneck. The real bottleneck was repeatedly copying a massive string just to search a few characters in it. Fixed by copying just the portion of the string we need to search. See flame graphs before and after.

![before](https://github.com/jenkinsci/json-lib/assets/29850/4ef3e19c-ea10-443b-9ba3-fb548dbe5623)

![after](https://github.com/jenkinsci/json-lib/assets/29850/639408aa-ff28-48fc-bda0-732c5cece146)

### Next steps

If merged and deployed to Jenkins users successfully, I will try to propose this upstream at https://github.com/kordamp/json-lib.

### Testing done

Tested against the real Jenkins Update Center JSON file; execution time went from about 8 seconds to about 0.4 seconds consistently. Tested interactively in Jenkins as well; the "Refresh" button on the plugin update page is now bounded by download time, not CPU time, and for me goes from about 12 seconds to about 5 seconds.