@jglick 2024-04-26
So https://github.com/jenkinsci/json-lib/compare/7a7f18e301f5571281f6843d10f935321c39ef6b...f5aa4ac42834433b920fa1534519ae7bbdd98c11 to avoid https://github.com/jenkinsci/json-lib/blob/f5aa4ac42834433b920fa1534519ae7bbdd98c11/src/main/java/net/sf/json/util/JSONTokener.java#L108-L111 which is going to be inefficient (especially if the document is not strictly ISO-8859-1 as in the case of the UC). Thanks for fixing this long-standing problem!